# functions/api/x402/v1/[[route]].ts
// EarlyThunder Research Data API - paid x402 endpoints (USDC on Base + Solana).
// Served as a Cloudflare Pages Function alongside the static site. Facilitator is
// keyless PayAI; payload snapshots are regenerated by scripts/build-x402-payloads.mjs
// at deploy time from the canonical research data.
import { Hono } from "hono";
import { handle } from "hono/cloudflare-pages";
import { paymentMiddleware, x402ResourceServer } from "@x402/hono";
import { ExactEvmScheme } from "@x402/evm/exact/server";
import { ExactSvmScheme } from "@x402/svm/exact/server";
import { HTTPFacilitatorClient } from "@x402/core/server";

import scores from "./_scores.json";
import theses from "./_theses.json";
import catalysts from "./_catalysts.json";
import scorecard from "./_scorecard.json";
import earnings from "./_earnings.json";

interface Env {
  PAY_TO_EVM: string;
  PAY_TO_SVM: string;
  FACILITATOR_URL: string;
}

const EVM_NETWORK = "eip155:8453"; // Base mainnet
const SVM_NETWORK = "solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp"; // Solana mainnet

const PRICES = {
  scores: "$0.01",
  catalysts: "$0.03",
  scorecard: "$0.05",
  earnings: "$0.01",
  thesis: "$0.25",
} as const;

function accepts(env: Env, price: string) {
  return [
    {
      scheme: "exact",
      network: EVM_NETWORK,
      payTo: env.PAY_TO_EVM,
      price,
      maxTimeoutSeconds: 300,
    },
    {
      scheme: "exact",
      network: SVM_NETWORK,
      payTo: env.PAY_TO_SVM,
      price,
      maxTimeoutSeconds: 300,
    },
  ];
}

const CATALOG = {
  service: "EarlyThunder Research Data API",
  site: "https://earlythunder.com",
  skill: "https://earlythunder.com/SKILL.md",
  openapi: "https://earlythunder.com/openapi.json",
  payment: "x402 (USDC on Base eip155:8453 or Solana mainnet), exact scheme. The live 402 challenge is authoritative.",
  endpoints: [
    { path: "/api/x402/v1/scores", price: PRICES.scores, description: "Ranked composite scores, tiers and one-liners for 259 crypto assets" },
    { path: "/api/x402/v1/thesis?slug={slug}", price: PRICES.thesis, description: "Full cited deep-value thesis for one asset (slug from /scores, e.g. ethereum)" },
    { path: "/api/x402/v1/catalysts", price: PRICES.catalysts, description: "Dated regulatory and protocol deadlines with urgency ratings and sources" },
    { path: "/api/x402/v1/scorecard", price: PRICES.scorecard, description: "Quant scorecard for 251 tokens with scoring methodology" },
    { path: "/api/x402/v1/earnings", price: PRICES.earnings, description: "Top protocol earnings yields with risk ratings" },
  ],
};

let cached: ReturnType<typeof handle> | null = null;

function buildApp(env: Env) {
  const facilitatorClient = new HTTPFacilitatorClient({ url: env.FACILITATOR_URL });
  const resourceServer = new x402ResourceServer(facilitatorClient)
    .register(EVM_NETWORK, new ExactEvmScheme())
    .register(SVM_NETWORK, new ExactSvmScheme());

  const app = new Hono();

  app.use("*", async (c, next) => {
    await next();
    c.header("Cache-Control", "no-store");
    c.header("Access-Control-Expose-Headers", "PAYMENT-REQUIRED, PAYMENT-RESPONSE, X-PAYMENT-RESPONSE");
  });

  app.options("*", (c) => {
    c.header("Access-Control-Allow-Origin", "*");
    c.header("Access-Control-Allow-Methods", "GET, OPTIONS");
    c.header("Access-Control-Allow-Headers", "Content-Type, PAYMENT-SIGNATURE, X-PAYMENT");
    return c.body(null, 204);
  });

  app.use(
    paymentMiddleware(
      {
        "GET /api/x402/v1/scores": {
          accepts: accepts(env, PRICES.scores),
          description:
            "Ranked crypto research scores from EarlyThunder: composite score, tier, category and one-line verdict for 259 assets, sorted best-first. Teaser for the full thesis endpoint.",
          mimeType: "application/json",
          serviceName: "EarlyThunder",
          tags: ["crypto", "research", "scores", "altcoins", "deep-value"],
          
        },
        "GET /api/x402/v1/thesis": {
          accepts: accepts(env, PRICES.thesis),
          description:
            "Full deep-value investment thesis for one crypto asset from EarlyThunder Research: composite score, tier, valuation reasoning and cited thesis text. Pass slug from the scores endpoint.",
          mimeType: "application/json",
          serviceName: "EarlyThunder",
          tags: ["crypto", "research", "thesis", "due-diligence", "deep-value"],
          
        },
        "GET /api/x402/v1/catalysts": {
          accepts: accepts(env, PRICES.catalysts),
          description:
            "Curated crypto catalyst and deadline calendar from EarlyThunder: dated regulatory and protocol deadlines with urgency ratings, status, notes and sources. Updated by an active research pipeline.",
          mimeType: "application/json",
          serviceName: "EarlyThunder",
          tags: ["crypto", "catalysts", "calendar", "deadlines", "regulation"],
          
        },
        "GET /api/x402/v1/scorecard": {
          accepts: accepts(env, PRICES.scorecard),
          description:
            "Quant altcoin scorecard from EarlyThunder: 251 tokens with price, market cap, ATH distance, TVL, circulating supply share, score, prior score and verdict, plus the scoring methodology.",
          mimeType: "application/json",
          serviceName: "EarlyThunder",
          tags: ["crypto", "scorecard", "screener", "altcoins", "quant"],
          
        },
        "GET /api/x402/v1/earnings": {
          accepts: accepts(env, PRICES.earnings),
          description:
            "Top crypto protocol earnings yields from EarlyThunder: yield percent, source of yield, TVL and risk rating for 15 protocols.",
          mimeType: "application/json",
          serviceName: "EarlyThunder",
          tags: ["crypto", "yield", "earnings", "defi", "staking"],
          
        },
      },
      resourceServer,
      undefined,
      undefined,
      true,
    ),
  );

  app.get("/api/x402/v1/index", (c) => c.json(CATALOG));
  app.get("/api/x402/v1/scores", (c) => c.json(scores));
  app.get("/api/x402/v1/catalysts", (c) => c.json(catalysts));
  app.get("/api/x402/v1/scorecard", (c) => c.json(scorecard));
  app.get("/api/x402/v1/earnings", (c) => c.json(earnings));
  app.get("/api/x402/v1/thesis", (c) => {
    const slug = (c.req.query("slug") || "").toLowerCase().trim();
    const bySlug = (theses as { bySlug: Record<string, unknown>; updated_at: string; source: string }).bySlug;
    const hit = bySlug[slug];
    if (!hit) {
      return c.json(
        {
          error: "unknown slug",
          hint: "Use a slug from /api/x402/v1/scores, e.g. ethereum",
        },
        404,
      );
    }
    return c.json({
      updated_at: (theses as { updated_at: string }).updated_at,
      source: (theses as { source: string }).source,
      ...(hit as Record<string, unknown>),
    });
  });

  return handle(app);
}

export const onRequest: PagesFunction<Env> = (ctx) => {
  if (!cached) cached = buildApp(ctx.env);
  return cached(ctx);
};